+++ packages/ai/src/tools/helpers.ts
